Add basic black/white drawing and reset functionality
Struggled to compartmentalize and nail down the coding mechanics for this project. After studying CSS grid, implemented the basics to get a variable resolution grid. Tried to figure out how to implement random color and greyscale, and they are proving difficult, though I understand the basics. Will keep working on those and update as progress is made.

Did finally come up with a good recursive function to check for an acceptable integer number of rows/columns, and placed it inside the resetGrid function.