Update script, solve functionality issues
Was confounded by why event listeners would carry over and cause functionality issues. Greyscale was the big culprit. Determining how to handle opacity (or as turned out to be better, alpha values) took up the majority of my study and brain space today. But I may have a working solution here. Need to add another else clause to handle circumstance where alphaValue is 1.0 and ends up reverting to 0.1. But that hopefully will not be too difficult.